Renamed some things and made them private
